Move files around a bit and build the executable
